+++ contracts/handlers/furucombo/IRCOMBO.sol
@@ -1,0 +1,6 @@
+pragma solidity ^0.6.0;
+
+interface IRCOMBO {
+    function provideFor(address, uint256) external;
+    function withdrawFor(address) external;
+}